Create Description and Findings
About the Problem and the Dataset:

In this assignment, there is an organization called as 'Cautious plc' that provides credit as its service offerings to its clients. It has recently compiled the data of its clients in order to build a model that could predict the clients, who could become defaulters or not pay the borrowed amount and interest on time. To be precise, the main agenda of Cautious plc is to utilize the data of its clients to predict the clients, who could end up not making the required payment of loans on time. Next task is to effectively reduce the number of features in the dataset for better understanding and accessibility. Cautious plc. intends to use the results of the analysis in this file to gain relevant insights about the clients, who could default on loan payments in the future and take necessary steps to avoid such circumstances in future.

Important Findings: 

1. Identified patterns in the data and relations between the features and the label by studying line plots, histograms, cross tables, correlation, and count plots.  

2. Compared the factors considered vital in the existing industry practices versus features in the dataset for classification of bad loans to infer the incompleteness of the given dataset and lack of proper justification for making the claims about default likelihood in the label of dataset. 

3. Made important observations about imbalanced misclassification costs attributed to incorrect prediction of bad versus good loans. Also tackled issues such as inconsistent information, high cardinality values, and choice of suitable metric for evaluation of classification. 

4. Employed a cost-sensitive learning approach that involved a combination of SMOTE (Synthetic Minority Oversampling Technique) to oversample from minority class (bad loans) and ENN (Edited Nearest Neighbors) to undersample from majority class (good loans) for building a classification model. Evaluated the model using F2 score metric for assigning greater weightage to recall or sensitivity than to precision for better prediction of bad loans.