Description
Structure of Files:

The data files that are given in this assignment are arranged as follows:

There are 8 users. For each of the 8 users, there are 20 gestures. There are 20 files for gesture 1 itself and last 3 columns in each of the files correspond to x,y, and z coordinates. It is also given that each of these 8 users repeat 20 times each of their respective 20 gestures. The code file contains code for data preparation for each of the 20 gestures repeated 20 times by each of the 8 users and these gestures are recorded by accelerometer sensors. This code file was run 8 times in which each time data from all the 20 text files that recorded the 20 instances of a gesture performed by a particular user was consolidated and converted into csv (comma separated value) files. Each of these csv files prepared after consolidation of data are also included here on the Github for reference. Basic ETL (Extract, Transform and Load) operations are performed here to read, consolidate and prepare the data from 3200 text files.