Basic ETL to analyse Accelerometer Sensor Data
The code file contains code for data preparation for each of the 20 gestures repeated 20 times by each of the 8 users and these gestures are recorded by accelerometer sensors. This code file was run 8 times in which each time data from all the 20 text files that recorded the 20 instances of a gesture performed by a particular user was consolidated and converted into csv (comma separated value) files. Each of these csv files prepared after consolidation of data are also included here on the Github for reference. Basic ETL (Extract, Transform and Load) operations are performed here to read, consolidate and prepare the data from 3200 text files.